runner.py/tracker.py: Add method to turn off header.
It seems logical to have method to turn off the test case header output
for users (like us) who don't want to see it. This commit adds an
additional optional parameter to the Tracker class to do so. The default
remains as True; the set_header method will allow for turning it
on or off.

This commit adds a method to the header class to turn this flag on/off
in the embedded _tracker object.

This class also adds tests for these new features (though it took some
work to get the tests to run. ;) )