Fix parsing of multi-line strings in YAML blocks
In 0aedb990c2 (#111) we tried to prevent extra whitespace from sneaking
in, but the chosen solution accidentally trimmed the last whitespace
character, leading to this YAML block:

  ok 1 /test
    ...
    x: |+
      text
    ---

Being incorrectly evaluated as {"x": "text"} instead of {"x": "text\n"}.

That's caused by the fact that we use `rstrip("\n")` to remove a newline
from each line, but paste them back together with "\n".join() which adds
a newline back to every line except the last.

Change the approach: instead of trying to strip newlines and add them
back again later, just leave them alone and join with the empty string.

Expand the test case for dealing with multiline strings to make sure we
get this right.